LPS-65012 Fix internal package in youtube
