+++ modules/apps/youtube/youtube-web/src/main/java/com/liferay/youtube/web/internal/constants/YouTubePortletKeys.java
@@ -12,21 +12,14 @@
  * details.
  */
 
-package com.liferay.youtube.web.portlet.action;
-
-import com.liferay.portal.kernel.portlet.ConfigurationAction;
-import com.liferay.portal.kernel.portlet.DefaultConfigurationAction;
-import com.liferay.youtube.web.constants.YouTubePortletKeys;
-
-import org.osgi.service.component.annotations.Component;
+package com.liferay.youtube.web.internal.constants;
 
 /**
  * @author Peter Fellwock
  */
-@Component(
-	immediate = true,
-	property = {"javax.portlet.name=" + YouTubePortletKeys.YOUTUBE},
-	service = ConfigurationAction.class
-)
-public class YouTubeConfigurationAction extends DefaultConfigurationAction {
+public class YouTubePortletKeys {
+
+	public static final String YOUTUBE =
+		"com_liferay_youtube_web_portlet_YouTubePortlet";
+
 }
